refactor: move helm values schema
